backend/requirements.txt
# FastAPI y servidor
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Validación y configuración
pydantic==2.8.2
pydantic-settings==2.4.0

# HTTP client
requests==2.32.3

# RAG y vectores
chromadb==0.5.4
sentence-transformers==2.7.0

# Procesamiento de documentos
pypdf==4.3.1
python-docx==1.1.0

# LLM providers (opcional, según configuración)
openai==1.12.0

# Evitar rompimientos por NumPy 2.x en algunos wheels
numpy<2.0
